+++ trade/src/main/java/bisq/trade/mu_sig/MuSigTradeService.java
@@ -59,8 +59,8 @@ import bisq.trade.mu_sig.messages.network.MuSigTradeMessage;
 import bisq.trade.mu_sig.protocol.MuSigBuyerAsTakerProtocol;
 import bisq.trade.mu_sig.protocol.MuSigProtocol;
 import bisq.trade.mu_sig.protocol.MuSigSellerAsMakerProtocol;
-import bisq.trade.mu_sig.protocol.ignore.MuSigBuyerAsMakerProtocol;
-import bisq.trade.mu_sig.protocol.ignore.MuSigSellerAsTakerProtocol;
+import bisq.trade.mu_sig.protocol.not_yet_impl.MuSigBuyerAsMakerProtocol;
+import bisq.trade.mu_sig.protocol.not_yet_impl.MuSigSellerAsTakerProtocol;
 import bisq.trade.protobuf.MusigGrpc;
 import bisq.trade.protobuf.SubscribeTxConfirmationStatusRequest;
 import bisq.user.banned.BannedUserService;
